# lib/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Api
  version: 1.0.0
  description: |
    Typed contract for the Complete iQ Carrier Audit API.

    Authenticated requests use the `sid` HTTP-only session cookie. The optional
    `X-Organization-Id` header selects one of the caller's organization
    memberships; otherwise the default (or earliest) membership is used.
    Organization-scoped lookups deliberately return `404` when a resource
    belongs to another tenant so that cross-tenant identifiers are not
    disclosed.

    Job-enqueueing operations that accept `X-Idempotency-Key` scope that value
    to the selected organization and operation inputs. Repeating an equivalent
    request returns the existing durable processing job and sets `duplicate`
    to `true` when that field is present.
servers:
  - url: /api
    description: Application API
security:
  - cookieAuth: []
...
paths:
...
  /auth/user:
    get:
      operationId: getAuthSession
      tags: [Authentication]
      summary: Get the current session
      description: "Returns `user: null` when no valid session cookie is present."
      security:
        - {}
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      responses:
        "200":
          description: Current user and selected organization context, if authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/login:
    post:
      operationId: login
      tags: [Authentication]
      summary: Create a session
      description: |
        Validates credentials and sets the HTTP-only `sid` cookie (`Secure`
        outside development). Failed sign-in attempts are limited to 10 per
        15 minutes; successful requests do not count against that limiter.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Credentials accepted and session created.
          headers:
            Set-Cookie:
              description: HTTP-only session cookie, marked Secure outside development.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/logout:
    post:
      operationId: logout
      tags: [Authentication]
      summary: Clear the current session
      description: Deletes the server-side session when present and clears the `sid` cookie.
      security:
        - {}
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      responses:
        "200":
          description: Session cleared or no session was present.
          headers:
            Set-Cookie:
              description: Expired session cookie.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid
      description: HTTP-only server session cookie, marked Secure outside development.
    inboundToken:
      type: apiKey
      in: query
      name: token
      description: Shared secret configured for the SendGrid Inbound Parse callback.
  parameters:
    OrganizationId:
      name: X-Organization-Id
      in: header
      required: false
      description: |
        Organization membership to use for this request. Omit to use the
        caller's default membership. An inaccessible value returns `403`.
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Request parameters or body failed route validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: A valid session or organization membership is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The caller lacks the selected organization or required permission.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    TooManyRequests:
      description: Request rate limit exceeded.
      headers:
        RateLimit-Limit:
          schema:
            type: integer
        RateLimit-Remaining:
          schema:
            type: integer
        RateLimit-Reset:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerError:
      description: The server could not complete the request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      description: Shared JSON error payload.
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
        code:
          type: string
          description: Optional stable machine-readable code.
        details:
          description: Optional structured diagnostic context.
          nullable: true
        requestId:
          type: string
          description: Optional request correlation identifier.
      additionalProperties: false
    Success:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
      additionalProperties: false
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
          format: password
      additionalProperties: false
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl, role]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        profileImageUrl:
          type: string
          nullable: true
        role:
          type: string
      additionalProperties: false
    OrganizationRole:
      type: string
      enum: [owner, admin, auditor, reviewer, member, viewer]
    OrganizationPermission:
      type: string
      enum:
        - claims:read
        - claims:create
        - claims:update
        - claims:delete
        - claims:assign
        - audits:run
        - findings:review
        - jobs:read
        - jobs:cancel
        - jobs:retry
        - views:manage
        - settings:manage
        - email:send
    AuthOrganization:
      type: object
      required: [id, name, role, permissions]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        role:
          $ref: "#/components/schemas/OrganizationRole"
        permissions:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/OrganizationPermission"
      additionalProperties: false
    AuthSession:
      type: object
      required: [user]
      properties:
        user:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/AuthUser"
        organization:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/AuthOrganization"
      additionalProperties: false
    LoginResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
      additionalProperties: false
...